Bad cowsay cow mech-and-cow was causing error
Fixed incorrect ls spec "*cow" to "*.cow" that was causing the error.
Now use _BAD_COWS array instead of multiple if statements.
Clean up passing in a cowfile on the command line, did not appear to
work correctly.
Clean up function get_cowfile to not sanitize the list but to do it
latter when on when using the returned cowfile.
Clean up incorrect name of file from fortune.sh to 10-header-fortune.sh