macOS added call to path_helper
During an update of macOS, the update added a call to path_helper to the bash_profile.
This created a problem because it prepends the paths to the PATH and pyenv needs to
be first (or before an actual python install) or it does not work correctly. I moved
the eval of path_helper to the start of the file and all is well now.